+++ examples/tutorial/basic-1/programs/basic-1/src/lib.rs
@@ -1,5 +1,7 @@
 use anchor_lang::prelude::*;
 
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
+
 #[program]
 mod basic_1 {
     use super::*;
@@ -20,7 +22,7 @@ mod basic_1 {
 #[derive(Accounts)]
 pub struct Initialize<'info> {
     #[account(init, payer = user, space = 8 + 8)]
-    pub my_account: ProgramAccount<'info, MyAccount>,
+    pub my_account: Account<'info, MyAccount>,
     pub user: AccountInfo<'info>,
     pub system_program: AccountInfo<'info>,
 }
@@ -28,7 +30,7 @@ pub struct Initialize<'info> {
 #[derive(Accounts)]
 pub struct Update<'info> {
     #[account(mut)]
-    pub my_account: ProgramAccount<'info, MyAccount>,
+    pub my_account: Account<'info, MyAccount>,
 }
 
 #[account]