+++ examples/tutorial/basic-2/programs/basic-2/src/lib.rs
@@ -1,7 +1,7 @@
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::system_program;
 
-// Define the program's instruction handlers.
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
 
 #[program]
 mod basic_2 {
@@ -21,12 +21,10 @@ mod basic_2 {
     }
 }
 
-// Define the validated accounts for each handler.
-
 #[derive(Accounts)]
 pub struct Create<'info> {
     #[account(init, payer = user, space = 8 + 40)]
-    pub counter: ProgramAccount<'info, Counter>,
+    pub counter: Account<'info, Counter>,
     #[account(signer)]
     pub user: AccountInfo<'info>,
     #[account(address = system_program::ID)]
@@ -36,12 +34,10 @@ pub struct Create<'info> {
 #[derive(Accounts)]
 pub struct Increment<'info> {
     #[account(mut, has_one = authority)]
-    pub counter: ProgramAccount<'info, Counter>,
+    pub counter: Account<'info, Counter>,
     #[account(signer)]
     pub authority: AccountInfo<'info>,
 }
-
-// Define the program owned accounts.
 
 #[account]
 pub struct Counter {