+++ tests/events/programs/events/src/lib.rs
@@ -3,6 +3,8 @@
 
 use anchor_lang::prelude::*;
 
+declare_id!("2dhGsWUzy5YKUsjZdLHLmkNpUDAXkNa9MYWsPc4Ziqzy");
+
 #[program]
 pub mod events {
     use super::*;