+++ tests/misc/programs/misc/src/lib.rs
@@ -10,6 +10,8 @@ mod account;
 mod context;
 mod event;
 
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
+
 #[program]
 pub mod misc {
     use super::*;