+++ tests/zero-copy/programs/zero-copy/src/lib.rs
@@ -7,6 +7,8 @@
 
 use anchor_lang::prelude::*;
 
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
+
 #[program]
 pub mod zero_copy {
     use super::*;